+++ sdk/resourcemanager/eventhub/armeventhub/zz_generated_pollers.go
@@ -229,6 +229,49 @@ func (p *NamespacesClientDeletePoller) ResumeToken() (string, error) {
 	return p.pt.ResumeToken()
 }
 
+// NetworkSecurityPerimeterConfigurationsClientCreateOrUpdatePoller provides polling facilities until the operation reaches a terminal state.
+type NetworkSecurityPerimeterConfigurationsClientCreateOrUpdatePoller struct {
+	pt *azcore.Poller
+}
+
+// Done returns true if the LRO has reached a terminal state.
+func (p *NetworkSecurityPerimeterConfigurationsClientCreateOrUpdatePoller) Done() bool {
+	return p.pt.Done()
+}
+
+// Poll fetches the latest state of the LRO.  It returns an HTTP response or error.
+// If the LRO has completed successfully, the poller's state is updated and the HTTP
+// response is returned.
+// If the LRO has completed with failure or was cancelled, the poller's state is
+// updated and the error is returned.
+// If the LRO has not reached a terminal state, the poller's state is updated and
+// the latest HTTP response is returned.
+// If Poll fails, the poller's state is unmodified and the error is returned.
+// Calling Poll on an LRO that has reached a terminal state will return the final
+// HTTP response or error.
+func (p *NetworkSecurityPerimeterConfigurationsClientCreateOrUpdatePoller) Poll(ctx context.Context) (*http.Response, error) {
+	return p.pt.Poll(ctx)
+}
+
+// FinalResponse performs a final GET to the service and returns the final response
+// for the polling operation. If there is an error performing the final GET then an error is returned.
+// If the final GET succeeded then the final NetworkSecurityPerimeterConfigurationsClientCreateOrUpdateResponse will be returned.
+func (p *NetworkSecurityPerimeterConfigurationsClientCreateOrUpdatePoller) FinalResponse(ctx context.Context) (NetworkSecurityPerimeterConfigurationsClientCreateOrUpdateResponse, error) {
+	respType := NetworkSecurityPerimeterConfigurationsClientCreateOrUpdateResponse{}
+	resp, err := p.pt.FinalResponse(ctx, nil)
+	if err != nil {
+		return NetworkSecurityPerimeterConfigurationsClientCreateOrUpdateResponse{}, err
+	}
+	respType.RawResponse = resp
+	return respType, nil
+}
+
+// ResumeToken returns a value representing the poller that can be used to resume
+// the LRO at a later time. ResumeTokens are unique per service operation.
+func (p *NetworkSecurityPerimeterConfigurationsClientCreateOrUpdatePoller) ResumeToken() (string, error) {
+	return p.pt.ResumeToken()
+}
+
 // PrivateEndpointConnectionsClientDeletePoller provides polling facilities until the operation reaches a terminal state.
 type PrivateEndpointConnectionsClientDeletePoller struct {
 	pt *azcore.Poller